src/main/java/com/travel/planner/utils/RoleMapper.java
package com.travel.planner.utils;

import com.travel.planner.dto.RoleDTO;
import com.travel.planner.entity.Role;
import org.springframework.stereotype.Component;

@Component
public class RoleMapper {
    public RoleDTO roleToDTO (Role role) {
        RoleDTO roleDTO = new RoleDTO();
        roleDTO.setId(role.getId());
        roleDTO.setName(role.getName());
        return roleDTO;
    }

    public Role roleToEntity (RoleDTO roleDTO) {
        Role role = new Role();
        role.setId(roleDTO.getId());
        role.setName(roleDTO.getName());
        return role;
    }
}
